requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
language-tool-python==2.7.1
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
openai==1.102.0
orjson==3.11.3
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyphen==0.17.2
python-docx==1.1.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
textstat==0.7.4
tqdm==4.67.1
typer==0.16.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
